+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 - add `stop_gradient` method to backends
 
+- add `has_aux` arg on `vvag` method
+
 ### Changed
 
 - contractor, dtype and backend set are default to return objects, `with tc.runtime_backend("jax") as K` or `K = tc.set_backend("jax")` could work